requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scapy==2.5.0
sqlalchemy==2.0.30
celery==5.4.0
redis==5.0.4
scikit-learn==1.5.0
tensorflow==2.16.1
numpy==1.26.4
joblib==1.4.2
python-nmap==0.7.1
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
